feat: select logic to query options
