+++ src/core_modules/capture-core/components/WidgetProfile/OverflowMenu/Delete/DeleteModal/hooks/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { useDeleteTrackedEntity } from './useDeleteTrackedEntity';
