+++ src/core_modules/capture-core/components/WidgetProfile/OverflowMenu/OverflowMenu.container.js
@@ -1,22 +1,11 @@
 // @flow
-import React, { useMemo } from 'react';
+import React from 'react';
 import type { Props } from './OverflowMenu.types';
 import { OverflowMenuComponet } from './OverflowMenu.component';
-import { useAuthorities } from './useAuthorities';
-
-const auth = {
-    F_TEI_CASCADE_DELETE: 'F_TEI_CASCADE_DELETE',
-    ALL: 'ALL',
-};
+import { useAuthorities } from './hooks';
 
 export const OverflowMenu = ({ trackedEntityTypeName, canWriteData, trackedEntity, onDeleteSuccess }: Props) => {
-    const { authorities } = useAuthorities();
-    const canCascadeDeleteTei = useMemo(
-        () =>
-            authorities &&
-            authorities.some(authority => authority === auth.ALL || authority === auth.F_TEI_CASCADE_DELETE),
-        [authorities],
-    );
+    const { canCascadeDeleteTei } = useAuthorities();
 
     return (
         <OverflowMenuComponet
