+++ src/core_modules/capture-core/components/WidgetProfile/OverflowMenu/hooks/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { useAuthorities } from './useAuthorities';
